For Windows Users
I used your project as a point of start and I was having problems with spawn when running on Windows. Then I used the same code on a Ubuntu VM and everything worked perfectly. So, after some search I discovered that windows users must pass  shell: true as an option to spawn. After adding it the error was gone and the code was running in both places.
Also, thank you for the code. It is helping me a lot to learn about sfdx plugins.